refactor: 커스텀 쉐이더에 대한 타입 지정
커스텀 쉐이더의 타입을 지정하고 컴포넌트 프롭에 대한 타입을 지정하여 any와 ts-ignore를 제거
